Split ACP client into transport and session-update layers
- adapters/acp/transport.rs: JSON-RPC 2.0 peer (RpcPeer) and the stdin
  read loop.
- adapters/acp/session_update_mapper.rs: pure mapping from session/update
  payloads to RunEvent or a Tool signal for stateful handling.
- adapters/acp/client.rs now focuses on AcpClient's handlers (permission
  flow, filesystem, terminal, stateful tool tracking) and delegates
  stateless mapping to the mapper and transport concerns to transport.rs.
