+++ src-tauri/src/adapters/acp/client.rs
@@ -2,117 +2,24 @@ use anyhow::{Context, Result, anyhow, bail};
 use serde_json::{Value, json};
 use std::{collections::HashMap, fs, path::PathBuf, process::Stdio, sync::Arc};
 use tokio::{
-    io::{AsyncBufReadExt, AsyncRead, AsyncReadExt, AsyncWriteExt, BufReader},
-    process::{Child, ChildStdin, Command},
-    sync::{Mutex, oneshot},
+    io::AsyncReadExt,
+    process::{Child, Command},
+    sync::Mutex,
 };
 use uuid::Uuid;
 
 use crate::{
-    adapters::acp::util::{
-        clean_tool_title, display_command, expand_tilde, extract_locations, normalize_path,
-        select_lines, string_param,
+    adapters::acp::{
+        session_update_mapper::{MappedSessionUpdate, map_session_update},
+        transport::RpcPeer,
+        util::{
+            clean_tool_title, display_command, expand_tilde, extract_locations, normalize_path,
+            select_lines, string_param,
+        },
     },
-    domain::events::{LifecycleStatus, PermissionOption, PlanEntry, RunEvent},
+    domain::events::{LifecycleStatus, PermissionOption, RunEvent},
     ports::{event_sink::RunEventSink, permission::PermissionDecisionPort},
 };
-
-use super::util::RpcError;
-
-type PendingMap = HashMap<u64, oneshot::Sender<std::result::Result<Value, RpcError>>>;
-
-#[derive(Clone)]
-pub struct RpcPeer {
-    writer: Arc<Mutex<ChildStdin>>,
-    pending: Arc<Mutex<PendingMap>>,
-    next_id: Arc<Mutex<u64>>,
-}
-
-impl RpcPeer {
-    pub fn new(stdin: ChildStdin) -> Self {
-        Self {
-            writer: Arc::new(Mutex::new(stdin)),
-            pending: Arc::new(Mutex::new(HashMap::new())),
-            next_id: Arc::new(Mutex::new(0)),
-        }
-    }
-
-    pub async fn request(
-        &self,
-        method: &str,
-        params: Value,
-    ) -> std::result::Result<Value, RpcError> {
-        let id = {
-            let mut next_id = self.next_id.lock().await;
-            let id = *next_id;
-            *next_id += 1;
-            id
-        };
-        let (tx, rx) = oneshot::channel();
-        self.pending.lock().await.insert(id, tx);
-        let payload = json!({"jsonrpc": "2.0", "id": id, "method": method, "params": params});
-        if let Err(err) = self.send_value(&payload).await {
-            let _ = self.pending.lock().await.remove(&id);
-            return Err(RpcError {
-                code: -32603,
-                message: err.to_string(),
-                data: None,
-            });
-        }
-        match rx.await {
-            Ok(result) => result,
-            Err(err) => Err(RpcError {
-                code: -32603,
-                message: format!("connection closed while waiting for {method}: {err}"),
-                data: None,
-            }),
-        }
-    }
-
-    async fn respond_ok(&self, id: Value, result: Value) -> Result<()> {
-        self.send_value(&json!({"jsonrpc": "2.0", "id": id, "result": result}))
-            .await
-    }
-
-    async fn respond_error(
-        &self,
-        id: Value,
-        code: i64,
-        message: &str,
-        data: Option<Value>,
-    ) -> Result<()> {
-        let mut error = json!({"code": code, "message": message});
-        if let Some(data) = data {
-            error["data"] = data;
-        }
-        self.send_value(&json!({"jsonrpc": "2.0", "id": id, "error": error}))
-            .await
-    }
-
-    async fn fail_pending(&self, message: impl Into<String>) {
-        let message = message.into();
-        let pending = {
-            let mut pending = self.pending.lock().await;
-            std::mem::take(&mut *pending)
-        };
-        for (_, tx) in pending {
-            let _ = tx.send(Err(RpcError {
-                code: -32603,
-                message: message.clone(),
-                data: None,
-            }));
-        }
-    }
-
-    async fn send_value(&self, value: &Value) -> Result<()> {
-        let mut writer = self.writer.lock().await;
-        let mut bytes = serde_json::to_vec(value)?;
-        bytes.push(b'\n');
-        writer.write_all(&bytes).await?;
-        writer.flush().await?;
-        Ok(())
-    }
-}
 
 struct TerminalState {
     child: Arc<Mutex<Child>>,
@@ -164,6 +71,12 @@ where
         self.sink.emit(&self.run_id, event);
     }
 
+    /// Emit a RunEvent on behalf of the transport layer (used from
+    /// `read_loop` to report JSON-RPC parse failures).
+    pub fn emit_raw(&self, event: RunEvent) {
+        self.emit(event);
+    }
+
     pub async fn handle_request(
         self: Arc<Self>,
         peer: RpcPeer,
@@ -210,11 +123,6 @@ where
     pub async fn handle_notification(&self, method: &str, params: Value) {
         if method == "session/update" {
             self.session_update(params).await;
-        } else if method.starts_with("ext/") {
-            self.emit(RunEvent::Raw {
-                method: method.to_string(),
-                payload: params,
-            });
         } else {
             self.emit(RunEvent::Raw {
                 method: method.to_string(),
@@ -305,68 +213,10 @@ where
     }
 
     async fn session_update(&self, params: Value) {
-        let Some(update) = params.get("update") else {
-            self.emit(RunEvent::Raw {
-                method: "session/update".into(),
-                payload: params,
-            });
-            return;
-        };
-        let kind = update
-            .get("sessionUpdate")
-            .and_then(Value::as_str)
-            .unwrap_or("session/update");
-        match kind {
-            "agent_message_chunk" => {
-                if let Some(text) = update.pointer("/content/text").and_then(Value::as_str) {
-                    self.emit(RunEvent::AgentMessage { text: text.into() });
-                }
-            }
-            "agent_thought_chunk" => {
-                if let Some(text) = update.pointer("/content/text").and_then(Value::as_str) {
-                    self.emit(RunEvent::Thought { text: text.into() });
-                }
-            }
-            "plan" => {
-                let entries = update
-                    .get("entries")
-                    .and_then(Value::as_array)
-                    .map(|entries| {
-                        entries
-                            .iter()
-                            .map(|entry| PlanEntry {
-                                status: entry
-                                    .get("status")
-                                    .and_then(Value::as_str)
-                                    .unwrap_or("")
-                                    .to_string(),
-                                content: entry
-                                    .get("content")
-                                    .and_then(Value::as_str)
-                                    .unwrap_or("")
-                                    .to_string(),
-                            })
-                            .collect::<Vec<_>>()
-                    })
-                    .unwrap_or_default();
-                self.emit(RunEvent::Plan { entries });
-            }
-            "tool_call" | "tool_call_update" => self.tool_update(update).await,
-            "usage_update" => {
-                let used = update
-                    .get("used")
-                    .and_then(Value::as_i64)
-                    .unwrap_or_default();
-                let size = update
-                    .get("size")
-                    .and_then(Value::as_i64)
-                    .unwrap_or_default();
-                self.emit(RunEvent::Usage { used, size });
-            }
-            other => self.emit(RunEvent::Raw {
-                method: other.to_string(),
-                payload: update.clone(),
-            }),
+        match map_session_update(&params) {
+            MappedSessionUpdate::Event(event) => self.emit(event),
+            MappedSessionUpdate::Tool(update) => self.tool_update(&update).await,
+            MappedSessionUpdate::Ignored => {}
         }
     }
 
@@ -636,82 +486,6 @@ where
     }
 }
 
-pub async fn read_loop<R, S, P>(
-    mut reader: BufReader<R>,
-    peer: RpcPeer,
-    client: Arc<AcpClient<S, P>>,
-    limit: usize,
-) -> Result<()>
-where
-    R: AsyncRead + Unpin,
-    S: RunEventSink,
-    P: PermissionDecisionPort,
-{
-    loop {
-        let mut bytes = Vec::new();
-        let read = reader.read_until(b'\n', &mut bytes).await?;
-        if read == 0 {
-            peer.fail_pending("ACP connection closed").await;
-            break;
-        }
-        if bytes.len() > limit {
-            peer.fail_pending("ACP message exceeded stdio buffer limit")
-                .await;
-            bail!("ACP message exceeded stdio buffer limit of {limit} bytes");
-        }
-        let message: Value = match serde_json::from_slice(&bytes) {
-            Ok(message) => message,
-            Err(err) => {
-                client.emit(RunEvent::Diagnostic {
-                    message: format!("failed to parse JSON-RPC message: {err}"),
-                });
-                continue;
-            }
-        };
-        let method = message
-            .get("method")
-            .and_then(Value::as_str)
-            .map(str::to_string);
-        let id = message.get("id").cloned();
-        match (method, id) {
-            (Some(method), Some(id)) => {
-                let params = message.get("params").cloned().unwrap_or(Value::Null);
-                let peer = peer.clone();
-                let client = Arc::clone(&client);
-                tokio::spawn(async move {
-                    client.handle_request(peer, id, method, params).await;
-                });
-            }
-            (Some(method), None) => {
-                let params = message.get("params").cloned().unwrap_or(Value::Null);
-                client.handle_notification(&method, params).await;
-            }
-            (None, Some(id)) => {
-                if let Some(request_id) = id.as_u64() {
-                    if let Some(tx) = peer.pending.lock().await.remove(&request_id) {
-                        if let Some(result) = message.get("result") {
-                            let _ = tx.send(Ok(result.clone()));
-                        } else {
-                            let error = message.get("error").cloned().unwrap_or(Value::Null);
-                            let _ = tx.send(Err(RpcError {
-                                code: error.get("code").and_then(Value::as_i64).unwrap_or(-32603),
-                                message: error
-                                    .get("message")
-                                    .and_then(Value::as_str)
-                                    .unwrap_or("Error")
-                                    .to_string(),
-                                data: error.get("data").cloned(),
-                            }));
-                        }
-                    }
-                }
-            }
-            (None, None) => {}
-        }
-    }
-    Ok(())
-}
-
 fn select_permission_option<'a>(options: &'a [Value], auto_allow: bool) -> Option<&'a Value> {
     if auto_allow {
         for desired in ["allow_once", "allow_always"] {
@@ -798,10 +572,8 @@ pub fn lifecycle(status: LifecycleStatus, message: impl Into<String>) -> RunEven
 
 #[cfg(test)]
 mod tests {
-    use super::{RpcPeer, select_permission_option};
+    use super::select_permission_option;
     use serde_json::json;
-    use std::process::Stdio;
-    use tokio::process::Command;
 
     #[test]
     fn auto_allow_prefers_allow_once_before_allow_always() {
@@ -831,25 +603,5 @@ mod tests {
             selected.get("optionId").and_then(|value| value.as_str()),
             Some("always")
         );
-    }
-
-    #[tokio::test]
-    async fn fail_pending_releases_waiting_requests() {
-        let mut child = Command::new("cat")
-            .stdin(Stdio::piped())
-            .stdout(Stdio::null())
-            .spawn()
-            .expect("spawn cat");
-        let stdin = child.stdin.take().expect("stdin");
-        let peer = RpcPeer::new(stdin);
-        let (tx, pending) = tokio::sync::oneshot::channel();
-        peer.pending.lock().await.insert(1, tx);
-
-        peer.fail_pending("closed for test").await;
-
-        let result = pending.await.expect("pending response");
-        assert!(result.is_err());
-        assert_eq!(result.err().expect("error").message, "closed for test");
-        let _ = child.kill().await;
     }
 }
