+++ src-tauri/src/adapters/acp/mod.rs
@@ -1,3 +1,5 @@
 pub mod client;
 pub mod runner;
+pub mod session_update_mapper;
+pub mod transport;
 pub mod util;
